+++ common/src/main/java/com/railwayteam/railways/mixin/NbtUtilsMixin.java
@@ -16,7 +16,8 @@
 
 package com.railwayteam.railways.mixin;
 
-import com.railwayteam.railways.base.datafixers.CompatCherryTrackFix;
+import com.railwayteam.railways.base.datafixerapi.DataFixesInternals;
+import com.railwayteam.railways.base.datafix.fixes.CompatCherryTrackFix;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtUtils;
@@ -32,6 +33,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(NbtUtils.class)
 public class NbtUtilsMixin {
     // moved general updating to MixinDataFixTypes.java
+    @Inject(method = "addDataVersion", at = @At("HEAD"))
+    private static void addModDataVersions(CompoundTag tag, int dataVersion, CallbackInfoReturnable<CompoundTag> cir) {
+        DataFixesInternals.get().addModDataVersions(tag);
+    }
 
     @Inject(method = "readBlockState", at = @At("HEAD"))
     private static void railways$upgradeBlocks(HolderGetter<Block> blockGetter, CompoundTag tag, CallbackInfoReturnable<BlockState> cir) {